fix(trackers): cast days pending to integer
- Added (int) cast to diffInDays() in both TrainingTracker and IdpTracker
- Prevents float values like 0.5099666278125 from displaying

+++ app/Livewire/Pages/Tracker/IdpTracker.php
@@ -234,7 +234,7 @@ class IdpTracker extends Component
       $planStatus = $this->getPlanStatusForUser($user);
 
       $daysPending = $planStatus['pending_since']
-        ? Carbon::parse($planStatus['pending_since'])->diffInDays(now())
+        ? (int) Carbon::parse($planStatus['pending_since'])->diffInDays(now())
         : null;
 
       return (object) [
